Fix flaky test
