+++ airbyte-workers/src/test/java/io/airbyte/workers/temporal/scheduling/ConnectionManagerWorkflowTest.java
@@ -48,13 +48,13 @@ import org.mockito.Mockito;
 
 public class ConnectionManagerWorkflowTest {
 
-  private static final ConfigFetchActivity mConfigFetchActivity =
+  private final ConfigFetchActivity mConfigFetchActivity =
       Mockito.mock(ConfigFetchActivity.class, Mockito.withSettings().withoutAnnotations());
-  private static final ConnectionDeletionActivity mConnectionDeletionActivity =
+  private final ConnectionDeletionActivity mConnectionDeletionActivity =
       Mockito.mock(ConnectionDeletionActivity.class, Mockito.withSettings().withoutAnnotations());
-  private static final GenerateInputActivityImpl mGenerateInputActivityImpl =
+  private final GenerateInputActivityImpl mGenerateInputActivityImpl =
       Mockito.mock(GenerateInputActivityImpl.class, Mockito.withSettings().withoutAnnotations());
-  private static final JobCreationAndStatusUpdateActivity mJobCreationAndStatusUpdateActivity =
+  private final JobCreationAndStatusUpdateActivity mJobCreationAndStatusUpdateActivity =
       Mockito.mock(JobCreationAndStatusUpdateActivity.class, Mockito.withSettings().withoutAnnotations());
 
   private TestWorkflowEnvironment testEnv;
@@ -511,143 +511,4 @@ public class ConnectionManagerWorkflowTest {
     }
 
   }
-
-  @Nested
-  @DisplayName("Test that sync workflow failures are recorded")
-  class SyncWorkflowReplicationFailuresRecorded {
-
-    private static final long JOB_ID = 111L;
-    private static final int ATTEMPT_ID = 222;
-
-    @BeforeEach
-    public void setup() {
-      testEnv = TestWorkflowEnvironment.newInstance();
-      worker = testEnv.newWorker(TemporalJobType.CONNECTION_UPDATER.name());
-      client = testEnv.getWorkflowClient();
-      worker.registerActivitiesImplementations(mConfigFetchActivity, mConnectionDeletionActivity, mGenerateInputActivityImpl,
-          mJobCreationAndStatusUpdateActivity);
-      workflow = client.newWorkflowStub(ConnectionManagerWorkflow.class,
-          WorkflowOptions.newBuilder().setTaskQueue(TemporalJobType.CONNECTION_UPDATER.name()).build());
-
-      Mockito.when(mConfigFetchActivity.getMaxAttempt()).thenReturn(new GetMaxAttemptOutput(1));
-    }
-
-    @Test
-    @DisplayName("Test that source and destination failures are recorded")
-    public void testSourceAndDestinationFailuresRecorded() {
-      worker.registerWorkflowImplementationTypes(ConnectionManagerWorkflowImpl.class, SourceAndDestinationFailureSyncWorkflow.class);
-      testEnv.start();
-
-      final UUID testId = UUID.randomUUID();
-      final TestStateListener testStateListener = new TestStateListener();
-      final WorkflowState workflowState = new WorkflowState(testId, testStateListener);
-      final ConnectionUpdaterInput input = new ConnectionUpdaterInput(UUID.randomUUID(), JOB_ID, ATTEMPT_ID, false, 1, workflowState, false);
-
-      WorkflowClient.start(workflow::run, input);
-      testEnv.sleep(Duration.ofMinutes(2L));
-      workflow.submitManualSync();
-
-      Mockito.verify(mJobCreationAndStatusUpdateActivity).attemptFailure(Mockito.argThat(new HasFailureFromSource(FailureOrigin.SOURCE)));
-      Mockito.verify(mJobCreationAndStatusUpdateActivity).attemptFailure(Mockito.argThat(new HasFailureFromSource(FailureOrigin.DESTINATION)));
-
-      testEnv.shutdown();
-    }
-
-    @Test
-    @DisplayName("Test that normalization failure is recorded")
-    public void testNormalizationFailure() {
-      worker.registerWorkflowImplementationTypes(ConnectionManagerWorkflowImpl.class, NormalizationFailureSyncWorkflow.class);
-      testEnv.start();
-
-      final UUID testId = UUID.randomUUID();
-      final TestStateListener testStateListener = new TestStateListener();
-      final WorkflowState workflowState = new WorkflowState(testId, testStateListener);
-      final ConnectionUpdaterInput input = new ConnectionUpdaterInput(UUID.randomUUID(), JOB_ID, ATTEMPT_ID, false, 1, workflowState, false);
-
-      WorkflowClient.start(workflow::run, input);
-      testEnv.sleep(Duration.ofMinutes(2L));
-      workflow.submitManualSync();
-
-      Mockito.verify(mJobCreationAndStatusUpdateActivity).attemptFailure(Mockito.argThat(new HasFailureFromSource(FailureOrigin.NORMALIZATION)));
-
-      testEnv.shutdown();
-    }
-
-    @Test
-    @DisplayName("Test that dbt failure is recorded")
-    public void testDbtFailureRecorded() {
-      worker.registerWorkflowImplementationTypes(ConnectionManagerWorkflowImpl.class, DbtFailureSyncWorkflow.class);
-      testEnv.start();
-
-      final UUID testId = UUID.randomUUID();
-      final TestStateListener testStateListener = new TestStateListener();
-      final WorkflowState workflowState = new WorkflowState(testId, testStateListener);
-      final ConnectionUpdaterInput input = new ConnectionUpdaterInput(UUID.randomUUID(), JOB_ID, ATTEMPT_ID, false, 1, workflowState, false);
-
-      WorkflowClient.start(workflow::run, input);
-      testEnv.sleep(Duration.ofMinutes(2L));
-      workflow.submitManualSync();
-
-      Mockito.verify(mJobCreationAndStatusUpdateActivity).attemptFailure(Mockito.argThat(new HasFailureFromSource(FailureOrigin.DBT)));
-
-      testEnv.shutdown();
-    }
-
-    @Test
-    @DisplayName("Test that persistence failure is recorded")
-    public void testPersistenceFailureRecorded() {
-      worker.registerWorkflowImplementationTypes(ConnectionManagerWorkflowImpl.class, PersistFailureSyncWorkflow.class);
-      testEnv.start();
-
-      final UUID testId = UUID.randomUUID();
-      final TestStateListener testStateListener = new TestStateListener();
-      final WorkflowState workflowState = new WorkflowState(testId, testStateListener);
-      final ConnectionUpdaterInput input = new ConnectionUpdaterInput(UUID.randomUUID(), JOB_ID, ATTEMPT_ID, false, 1, workflowState, false);
-
-      WorkflowClient.start(workflow::run, input);
-      testEnv.sleep(Duration.ofMinutes(2L));
-      workflow.submitManualSync();
-
-      Mockito.verify(mJobCreationAndStatusUpdateActivity).attemptFailure(Mockito.argThat(new HasFailureFromSource(FailureOrigin.PERSISTENCE)));
-
-      testEnv.shutdown();
-    }
-
-    @Test
-    @DisplayName("Test that replication worker failure is recorded")
-    public void testReplicationFailureRecorded() {
-      worker.registerWorkflowImplementationTypes(ConnectionManagerWorkflowImpl.class, ReplicateFailureSyncWorkflow.class);
-      testEnv.start();
-
-      final UUID testId = UUID.randomUUID();
-      final TestStateListener testStateListener = new TestStateListener();
-      final WorkflowState workflowState = new WorkflowState(testId, testStateListener);
-      final ConnectionUpdaterInput input = new ConnectionUpdaterInput(UUID.randomUUID(), JOB_ID, ATTEMPT_ID, false, 1, workflowState, false);
-
-      WorkflowClient.start(workflow::run, input);
-      testEnv.sleep(Duration.ofMinutes(2L));
-      workflow.submitManualSync();
-
-      Mockito.verify(mJobCreationAndStatusUpdateActivity).attemptFailure(Mockito.argThat(new HasFailureFromSource(FailureOrigin.REPLICATION_WORKER)));
-
-      testEnv.shutdown();
-    }
-
-  }
-
-  private class HasFailureFromSource implements ArgumentMatcher<AttemptFailureInput> {
-
-    private final FailureOrigin expectedFailureOrigin;
-
-    public HasFailureFromSource(final FailureOrigin failureOrigin) {
-      this.expectedFailureOrigin = failureOrigin;
-    }
-
-    @Override
-    public boolean matches(final AttemptFailureInput arg) {
-      return arg.getAttemptFailureSummary().getFailures().stream().anyMatch(f -> f.getFailureOrigin().equals(expectedFailureOrigin));
-    }
-
-  }
-
 }
